requirements.txt
tweepy
Pillow
schedule
boto3
python-dotenv
stability-sdk